# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Comics documentation",
        "description": "V1 Api docs",
        "contact": {
            "email": ""
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/media/{media_model}/{id}": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Список Изображений модели",
                "description": "Список Изображений модели",
                "operationId": "645112afee56a713e5b428b7b2c9ef2f",
                "parameters": [
                    {
                        "name": "media_model",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "manga",
                                "chapter"
                            ]
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MediaResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Media"
                ],
                "summary": "Добавить Изображение в модель",
                "description": "Добавить Изображение в модель",
                "operationId": "a54d24787ed883917fa24a7f86443f57",
                "parameters": [
                    {
                        "name": "media_model",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "manga",
                                "chapter"
                            ]
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HasMediaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ]
            }
        },
        "/media/{id}": {
            "delete": {
                "tags": [
                    "Media"
                ],
                "summary": "Удалить изображение",
                "description": "Удалить изображение",
                "operationId": "69ef316f361a35f4214e95e99d618256",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Media": {
                "title": "Media",
                "description": "Media Model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Идентификатор",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Название",
                        "type": "string",
                        "example": "Владимир"
                    },
                    "collection_name": {
                        "title": "collection_name",
                        "description": "collection name",
                        "type": "string",
                        "example": "default"
                    },
                    "disk": {
                        "title": "disk",
                        "description": "disk",
                        "type": "string",
                        "example": "public"
                    },
                    "url": {
                        "title": "url",
                        "description": "url",
                        "type": "string",
                        "example": "http://localhost"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Создана в",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Обновлена в",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Media"
                }
            },
...
            "HasMediaRequest": {
                "title": "Game request",
                "description": "Game request body data",
                "required": [],
                "properties": {
                    "media": {
                        "title": "media",
                        "description": "media",
                        "type": "array",
                        "items": {
                            "type": "file"
                        }
                    }
                },
                "type": "object"
            },
...
            "MediaResource": {
                "title": "MediaResource",
                "description": "Media Resource",
                "properties": {
                    "media": {
                        "$ref": "#/components/schemas/Media"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Media Resource"
                }
            },
...
        },
        "securitySchemes": {
            "api-key": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}